BezCode/frontend/requirements.txt
streamlit==1.50.0
paho-mqtt==1.6.1
plotly==6.0.0
pandas==2.3.3
streamlit-autorefresh==1.0.1